+++ plugins/org.jkiss.dbeaver.ext.db2.ui/OSGI-INF/l10n/bundle.properties
@@ -9,6 +9,8 @@ dialog.connection.header=DB2 Connection Settings
 
 editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Application Manager
 
+editor.source.declaration.name=Source
+editor.source.declaration.description=Source
 editor.source.ddl.name=DDL
 editor.source.ddl.description=DDL
