+++ plugins/org.jkiss.dbeaver.ext.db2.ui/OSGI-INF/l10n/bundle_de.properties
@@ -1,0 +1,21 @@
+#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
+
+category.db2.description = DB2 Werkzeuge
+category.db2.name        = DB2
+
+dialog.connection.header = DB2 Verbindungseinstellungen
+
+editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Applikationsmanager
+editor.source.ddl.description                                              = DDL
+editor.source.ddl.name                                                     = DDL
+editor.source.declaration.description                                      = Quelle
+editor.source.declaration.name                                             = Quelle
+
+tool.org.jkiss.dbeaver.ext.db2.table.reorg.name=Reorg Table...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgix.name=Reorg Index...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgcheck.name=Reorg Check (Table)...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgcheckix.name=Reorg Check (Indexes)...
+tool.org.jkiss.dbeaver.ext.db2.table.setintegrity.name=Set Integrity...
+tool.org.jkiss.dbeaver.ext.db2.table.runstats.name=Runstats...
+tool.org.jkiss.dbeaver.ext.db2.table.truncate.name=Truncate...
+tool.org.jkiss.dbeaver.ext.db2.showerror.name=Retrieve SQL Error Message...
