+++ plugins/org.jkiss.dbeaver.ext.db2.ui/OSGI-INF/l10n/bundle_fr.properties
@@ -1,0 +1,23 @@
+# DBeaver - Universal Database Manager
+# Copyright (C) 2013-2016 Denis Forveille titou10.titou10@gmail.com
+# Copyright (C) 2010-2020 DBeaver Corp and others serge@jkiss.org
+
+Bundle-Name = Support DB2 pour DBeaver
+
+tool.org.jkiss.dbeaver.ext.db2.table.reorg.name=R\u00E9organiser la table...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgix.name=R\u00E9organiser 'index...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgcheck.name=R\u00E9organiser le Check (Table)...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgcheckix.name=R\u00E9organiser le Check (Indexes)...
+tool.org.jkiss.dbeaver.ext.db2.table.setintegrity.name=D\u00E9finir l'int\u00E9grit\u00E9...
+tool.org.jkiss.dbeaver.ext.db2.table.runstats.name=Runstats...
+tool.org.jkiss.dbeaver.ext.db2.table.truncate.name=Vider...
+tool.org.jkiss.dbeaver.ext.db2.showerror.name=R\u00E9cup\u00E9rer le message d'erreur SQL...
+
+dialog.connection.header=Param\u00E8tres de connexion DB2
+
+editor.source.declaration.name=Source
+editor.source.declaration.description=Source
+editor.source.ddl.name=DDL
+editor.source.ddl.description=DDL
+
+editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Gestionnaire d'application
