+++ plugins/org.jkiss.dbeaver.ext.db2.ui/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -1,0 +1,24 @@
+#Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/) 
+# DBeaver - Universal Database Manager
+# Copyright (C) 2013-2016 Denis Forveille titou10.titou10@gmail.com
+# Copyright (C) 2010-2020 DBeaver Corp and others serge@jkiss.org
+
+Bundle-Name = DBeaver DB2 Support
+Bundle-Vendor = JKISS
+
+dialog.connection.header = DB2 Configura\u00E7\u00F5es de conex\u00E3o
+
+editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Application Manager
+editor.source.ddl.description                                              = DDL
+editor.source.ddl.name                                                     = DDL
+editor.source.declaration.description                                      = Origem
+editor.source.declaration.name                                             = Origem
+
+tool.org.jkiss.dbeaver.ext.db2.showerror.name          = Recuperar erros de messagem do SQL...
+tool.org.jkiss.dbeaver.ext.db2.table.reorg.name        = Reorganizar Tabela...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgcheck.name   = Reorganizar e Checar  (Tabela)...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgcheckix.name = Reorganizar Checar (Indices)...
+tool.org.jkiss.dbeaver.ext.db2.table.reorgix.name      = Reorganizar Indice...
+tool.org.jkiss.dbeaver.ext.db2.table.runstats.name     = Runstats...
+tool.org.jkiss.dbeaver.ext.db2.table.setintegrity.name = Setar Integridade...
+tool.org.jkiss.dbeaver.ext.db2.table.truncate.name     = Truncar...
