+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources.properties
@@ -82,3 +82,7 @@ dialog_tools_msg_code=SQL Code?
 dialog_tools_mes_error_code_title=Error
 dialog_tools_mes_error_code=The SQL Code must be Numeric, eg -551, 803...
 dialog_tools_mes_message=SQL Message
+
+edit_db2_constraint_manager_dialog_title=Create constraint
+edit_db2_foreign_key_manager_dialog_title=Create foreign key
+edit_db2_index_manager_dialog_title=Create index
