+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources_de.properties
@@ -61,3 +61,7 @@ dialog_tools_mes_error_code_title = Fehler
 dialog_tools_mes_message = SQL-Nachricht
 dialog_tools_msg_code = SQL-Code?
 dialog_tools_msg_title = Abrufen der SQL-Nachricht aus dem SQL-Code
+
+edit_db2_constraint_manager_dialog_title = Einschr\u00E4nkung erstellen
+edit_db2_foreign_key_manager_dialog_title = Fremdschl\u00FCssel erstellen
+edit_db2_index_manager_dialog_title = Index erstellen
