+++ plugins/org.jkiss.dbeaver.ext.db2.ui/src/org/jkiss/dbeaver/ext/db2/ui/internal/DB2Resources_ru.properties
@@ -69,8 +69,6 @@ dialog_table_tools_reorgix_options_cleanup_pages=Only Cleanup Pseudo Deleted Pag
 
 dialog_table_tools_reorgcheck_title=Reorg Check on Table
 
-dialog_table_tools_reorgcheckix_title=Reorg Check on Table Indexes
-
 dialog_table_tools_truncate_title=Truncate Table(s)
 dialog_table_tools_truncate_storage_title=Storage Options
 dialog_table_tools_truncate_storage_reuse=Reuse Storage
@@ -78,3 +76,6 @@ dialog_table_tools_truncate_storage_drop=Drop Storage
 dialog_table_tools_truncate_triggers_title=Delete Triggers Options 
 dialog_table_tools_truncate_triggers_ignore=Ignore Delete Triggers
 dialog_table_tools_truncate_triggers_restrict=Restrict When Delete Triggers
+
+edit_db2_foreign_key_manager_dialog_title=\u0421\u043E\u0437\u0434\u0430\u0442\u044C \u0432\u043D\u0435\u0448\u043D\u0438\u0439 \u043A\u043B\u044E\u0447
+edit_db2_index_manager_dialog_title=\u0421\u043E\u0437\u0434\u0430\u0442\u044C \u0438\u043D\u0434\u0435\u043A\u0441
