+++ plugins/org.jkiss.dbeaver.ext.db2/OSGI-INF/l10n/bundle.properties
@@ -8,22 +8,6 @@ Bundle-Name = DBeaver DB2 Support
 category.db2.description=DB2 Tools
 category.db2.name=DB2
 
-tool.org.jkiss.dbeaver.ext.db2.table.reorg.name=Reorg Table...
-tool.org.jkiss.dbeaver.ext.db2.table.reorgix.name=Reorg Index...
-tool.org.jkiss.dbeaver.ext.db2.table.reorgcheck.name=Reorg Check (Table)...
-tool.org.jkiss.dbeaver.ext.db2.table.reorgcheckix.name=Reorg Check (Indexes)...
-tool.org.jkiss.dbeaver.ext.db2.table.setintegrity.name=Set Integrity...
-tool.org.jkiss.dbeaver.ext.db2.table.runstats.name=Runstats...
-tool.org.jkiss.dbeaver.ext.db2.table.truncate.name=Truncate...
-tool.org.jkiss.dbeaver.ext.db2.showerror.name=Retrieve SQL Error Message...
-
-dialog.connection.header=DB2 Connection Settings
-
-editor.source.declaration.name=Source
-editor.source.declaration.description=Source
-editor.source.ddl.name=DDL
-editor.source.ddl.description=DDL
-
 menu.tables.manage=Manage
 
 tree.administer.node.name=Administer
@@ -864,5 +848,3 @@ meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleAuth.role.name=Role
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.admin.name=Admin?
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.grantee.name=Grantee
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.granteeType.name=Grantee Type
-
-editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Application Manager
