+++ plugins/org.jkiss.dbeaver.ext.db2/OSGI-INF/l10n/bundle_de.properties
@@ -3,14 +3,6 @@
 category.db2.description = DB2 Werkzeuge
 category.db2.name        = DB2
 
-dialog.connection.header = DB2 Verbindungseinstellungen
-
-editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Applikationsmanager
-editor.source.ddl.description                                              = DDL
-editor.source.ddl.name                                                     = DDL
-editor.source.declaration.description                                      = Quelle
-editor.source.declaration.name                                             = Quelle
-
 menu.tables.manage = Verwalten
 
 meta.org.jkiss.dbeaver.ext.db2.info.DB2Parameter.dataType.name                  = Typ
