+++ plugins/org.jkiss.dbeaver.ext.db2/OSGI-INF/l10n/bundle_fr.properties
@@ -7,22 +7,6 @@ Bundle-Name = Support DB2 pour DBeaver
 category.db2.description=Outils DB2
 category.db2.name=DB2
 
-tool.org.jkiss.dbeaver.ext.db2.table.reorg.name=R\u00E9organiser la table...
-tool.org.jkiss.dbeaver.ext.db2.table.reorgix.name=R\u00E9organiser 'index...
-tool.org.jkiss.dbeaver.ext.db2.table.reorgcheck.name=R\u00E9organiser le Check (Table)...
-tool.org.jkiss.dbeaver.ext.db2.table.reorgcheckix.name=R\u00E9organiser le Check (Indexes)...
-tool.org.jkiss.dbeaver.ext.db2.table.setintegrity.name=D\u00E9finir l'int\u00E9grit\u00E9...
-tool.org.jkiss.dbeaver.ext.db2.table.runstats.name=Runstats...
-tool.org.jkiss.dbeaver.ext.db2.table.truncate.name=Vider...
-tool.org.jkiss.dbeaver.ext.db2.showerror.name=R\u00E9cup\u00E9rer le message d'erreur SQL...
-
-dialog.connection.header=Param\u00E8tres de connexion DB2
-
-editor.source.declaration.name=Source
-editor.source.declaration.description=Source
-editor.source.ddl.name=DDL
-editor.source.ddl.description=DDL
-
 menu.tables.manage=G\u00E9rer
 
 tree.administer.node.name=Administrer
@@ -860,5 +844,3 @@ meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleAuth.role.name=R\u00F4le
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.admin.name=Admin?
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.grantee.name=receveur (Grantee)
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.granteeType.name=Type de receveur
-
-editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Gestionnaire d'application
