+++ plugins/org.jkiss.dbeaver.ext.db2/OSGI-INF/l10n/bundle_pt_BR.properties
@@ -10,14 +10,6 @@ Bundle-Vendor = JKISS
 category.db2.description = DB2 Tools
 category.db2.name        = DB2
 
-dialog.connection.header = DB2 Configura\u00E7\u00F5es de conex\u00E3o
-
-editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Application Manager
-editor.source.ddl.description                                              = DDL
-editor.source.ddl.name                                                     = DDL
-editor.source.declaration.description                                      = Origem
-editor.source.declaration.name                                             = Origem
-
 menu.tables.manage = Gerencia
 
 meta.org.jkiss.dbeaver.ext.db2.info.DB2Parameter.dataType.name                                = Tipo
@@ -644,15 +636,6 @@ meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleAuth.role.name             
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.admin.name                           = Admin?
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.grantee.name                         = Concedente
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.granteeType.name                     = Tipo de concedente
-
-tool.org.jkiss.dbeaver.ext.db2.showerror.name          = Recuperar erros de messagem do SQL...
-tool.org.jkiss.dbeaver.ext.db2.table.reorg.name        = Reorganizar Tabela...
-tool.org.jkiss.dbeaver.ext.db2.table.reorgcheck.name   = Reorganizar e Checar  (Tabela)...
-tool.org.jkiss.dbeaver.ext.db2.table.reorgcheckix.name = Reorganizar Checar (Indices)...
-tool.org.jkiss.dbeaver.ext.db2.table.reorgix.name      = Reorganizar Indice...
-tool.org.jkiss.dbeaver.ext.db2.table.runstats.name     = Runstats...
-tool.org.jkiss.dbeaver.ext.db2.table.setintegrity.name = Setar Integridade...
-tool.org.jkiss.dbeaver.ext.db2.table.truncate.name     = Truncar...
 
 tree.administer.node.name               = Administrador
 tree.alias.node.name                    = Apelido
