+++ plugins/org.jkiss.dbeaver.ext.db2/OSGI-INF/l10n/bundle_ru.properties
@@ -861,5 +861,3 @@ meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleAuth.role.name=Role
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.admin.name=Admin?
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.grantee.name=Grantee
 meta.org.jkiss.dbeaver.ext.db2.model.security.DB2RoleDep.granteeType.name=Grantee Type
-
-editor.org.jkiss.dbeaver.ext.db2.model.app.DB2ServerApplicationEditor.name = Application Manager
