+++ plugins/org.jkiss.dbeaver.ext.db2/src/org/jkiss/dbeaver/ext/db2/DB2Resources_de.properties
@@ -1,221 +1,17 @@
 #Generated by ResourceBundle Editor (http://essiembre.github.io/eclipse-rbe/)
-
-dialog_connection_advanced_tab = Erweitert
-
-dialog_connection_advanced_tab_tooltip = Erweiterte/angepasste Treibereigenschaften
-
-dialog_connection_basic_tab = Basis
-
-dialog_connection_browse_button = Durchsuchen ...
-
-dialog_connection_connection_type_group = Verbindungstyp
-
-dialog_connection_custom_tab = Benutzerdefiniert
-
-dialog_connection_database = Datenbank
-
-dialog_connection_database_schema_label = Datenbank/Schema:
-
-dialog_connection_db_file_chooser_text = Datenbankdatei w\u00E4hlen
-
-dialog_connection_db_folder_chooser_message = Ordner mit Datenbankdateien ausw\u00E4hlen
-
-dialog_connection_db_folder_chooser_text = Datenbankordner ausw\u00E4hlen
-
-dialog_connection_edit_driver_button = Treibereinstellungen bearbeiten
-
-dialog_connection_general_tab = Allgemein
-
-dialog_connection_general_tab_tooltip = Allgemeine Verbindungseigenschaften
-
-dialog_connection_host = Host
-
-dialog_connection_host_label = Host:
-
-dialog_connection_jdbc_url_ = JDBC URL:
-
-dialog_connection_ora_home = Client
-
-dialog_connection_os_authentication = Betriebssystem-Authentifizierung
-
-dialog_connection_password = Passwort
-
-dialog_connection_password_label = Passwort:
-
-dialog_connection_path_label = Pfad:
-
-dialog_connection_port = Port
-
-dialog_connection_port_label = Port:
-
-dialog_connection_role = Rolle
-
-dialog_connection_security_group = Sicherheit
-
-dialog_connection_select_ora_home_msg = DB2 Home ausw\u00E4hlen
-
-dialog_connection_server_label = Server:
-
-dialog_connection_service = Servicename
-
-dialog_connection_sid = SID
-
-dialog_connection_sid_service = SID/Service
-
-dialog_connection_test_connection = Verbindung testen ...
-
-dialog_connection_test_connection_button = Verbindung testen ...
-
-dialog_connection_tns_tab = TNS
-
-dialog_connection_user_name = Benutzername
-
-dialog_connection_user_name_label = Benutzername:
-
-dialog_connection_ver = v.
-
 dialog_explain_ask_to_create = Tabellen f\u00FCr EXPLAIN, die weder in der aktuellen Berechtigungs-ID (%s) noch in SYSTOOLS enthalten sind.\r\n\r\nM\u00F6chten Sie, dass DBeaver neue EXPLAIN-Tabellen erstellt?
-
 dialog_explain_choose_tablespace = Einen Tablespace ausw\u00E4hlen
-
 dialog_explain_choose_tablespace_tablespace = Tablespace
-
 dialog_explain_no_tables = Fehlende Tabellen erkl\u00E4ren
-
 dialog_explain_no_tables_found_ex = EXPLAIN-Tabellen nicht gefunden. Die Abfrage kann nicht beantwortet werden.
-
 dialog_explain_no_tablespace_found = Es wurde kein verwendbarer Tablespace gefunden, um Explain-Tabellen zu erstellen.
-
 dialog_explain_no_tablespace_found_title = Kein Tablespace
 
-dialog_table_tools_options = Optionen
-
-dialog_table_tools_progress = Fortschritt
-
-dialog_table_tools_reorg_access_no = Kein Zugriff erlauben
-
-dialog_table_tools_reorg_access_read = Nur Lesezugriff erlauben
-
-dialog_table_tools_reorg_access_readwrite = Lese- und Schreibzugriff erlauben
-
-dialog_table_tools_reorg_access_title = Tabellenzugriff
-
-dialog_table_tools_reorg_inplace = Reorg in Position?
-
-dialog_table_tools_reorg_reorglobs = Reorg LOBs und langen Spalten?
-
-dialog_table_tools_reorg_reorglobsts = Temp. Tablespace f\u00FCr LOBs und Lange Spalten verwenden
-
-dialog_table_tools_reorg_resetdict = W\u00F6rterbuch zur\u00FCcksetzen?
-
-dialog_table_tools_reorg_title = Reorg Tabelle
-
-dialog_table_tools_reorg_truncate = Tabelle nach Reorg abschneiden?
-
-dialog_table_tools_reorg_useindex = Reorg mit bestehendem Index?
-
-dialog_table_tools_reorg_useindexscan = Index-Scan verwenden?
-
-dialog_table_tools_reorg_usetempts = Tempor\u00E4ren Tablespace verwenden?
-
-dialog_table_tools_reorgcheck_title = Reorg-Pr\u00FCfung auf Tabelle
-
-dialog_table_tools_reorgcheckix_title = Reorg-Check auf Tabellenindizes
-
-dialog_table_tools_reorgix_access_default = Standardzugang verwenden
-
-dialog_table_tools_reorgix_access_no = Kein Zugriff erlauben
-
-dialog_table_tools_reorgix_access_read = Lesezugriff erlauben
-
-dialog_table_tools_reorgix_access_readwrite = Lese- und Schreibzugriff erlauben
-
-dialog_table_tools_reorgix_access_title = Tabellenzugriff
-
-dialog_table_tools_reorgix_options_cleanup_keys = Nur Pseudogel\u00F6schte Schl\u00FCssel und Seiten bereinigen
-
-dialog_table_tools_reorgix_options_cleanup_pages = Nur Pseudogel\u00F6schte Seiten bereinigen
-
-dialog_table_tools_reorgix_options_full = Vollst\u00E4ndiger Index Reorg
-
-dialog_table_tools_reorgix_options_title = Optionen
-
-dialog_table_tools_reorgix_title = Reorg Index auf Tabelle
-
-dialog_table_tools_result = Ergebnis
-
-dialog_table_tools_runstats_cols_all = Alle Spalten
-
-dialog_table_tools_runstats_cols_all_and_distribution = Alle Spalten mit Verteilung 
-
-dialog_table_tools_runstats_cols_no = Keine Statistiken zu den Spalten
-
-dialog_table_tools_runstats_cols_title = Sammeln Sie Kolumnenstatistiken
-
-dialog_table_tools_runstats_indexes_all = Alle Indizes
-
-dialog_table_tools_runstats_indexes_detailed = Erweiterte detaillierte Statistiken mit Stichproben auf allen Indizes
-
-dialog_table_tools_runstats_indexes_no = Keine Statistiken f\u00FCr Indizes
-
-dialog_table_tools_runstats_indexes_title = Sammeln von Indexstatistiken
-
-dialog_table_tools_runstats_stats_title = Tabellen-Stichproben (% der Seiten)
-
-dialog_table_tools_runstats_title = Runstats auf Tabelle
-
-dialog_table_tools_success_title = Erfolg
-
-dialog_table_tools_truncate_storage_drop = Speicher auswerfen
-
-dialog_table_tools_truncate_storage_reuse = Speicher wiederverwenden
-
-dialog_table_tools_truncate_storage_title = Speicheroptionen
-
-dialog_table_tools_truncate_title = Tabelle(n) abschneiden
-
-dialog_table_tools_truncate_triggers_ignore = L\u00F6schausl\u00F6ser ignorieren
-
-dialog_table_tools_truncate_triggers_restrict = Einschr\u00E4nken beim L\u00F6schen von Triggern
-
-dialog_table_tools_truncate_triggers_title = Trigger-Optionen l\u00F6schen 
-
-dialog_tools_mes_error_code = Der SQL-Code muss numerisch sein, z.B. -551, 803...
-
-dialog_tools_mes_error_code_title = Fehler
-
-dialog_tools_mes_message = SQL-Nachricht
-
-dialog_tools_msg_code = SQL-Code?
-
-dialog_tools_msg_title = Abrufen der SQL-Nachricht aus dem SQL-Code
-
-edit_db2_constraint_manager_dialog_title = Einschr\u00E4nkung erstellen
-
-edit_db2_data_type_manager_dialog_title = Paket
-
-edit_db2_foreign_key_manager_dialog_title = Fremdschl\u00FCssel erstellen
-
-edit_db2_index_manager_dialog_title = Index erstellen
-
-edit_db2_package_manager_dialog_title = Paket
-
-edit_db2_schema_manager_dialog_title = Schema name
-
-edit_db2_sequence_manager_dialog_title = Reihenfolge
-
-edit_db2_trigger_manager_dialog_title = Ausl\u00F6ser
-
 editors_db2_application_editor_action_force = Erzwingen
-
 editors_db2_application_editor_confirm_action = Sind Sie sicher, dass Sie die Anwendung "{1}" "erzwingen" wollen?
-
 editors_db2_application_editor_title_force_application = Erzwingen der Anwendung
 
 no_ddl_for_nicknames = -- DDL ist f\u00FCr Nicknames nicht verf\u00FCgbar.
-
 no_ddl_for_nonsql_routines = -- DDL ist nicht verf\u00FCgbar f\u00FCr nicht-SQL-basierte Routinen.
-
 no_ddl_for_spaces_in_name = -- DDL ist nicht verf\u00FCgbar f\u00FCr Objekte, die einen Whitespace in ihrem Namen oder Schema haben.
-
 no_ddl_for_system_tables = -- DDL ist f\u00FCr Systemtabellen nicht verf\u00FCgbar.
