+++ plugins/org.jkiss.dbeaver.ext.db2/src/org/jkiss/dbeaver/ext/db2/manager/DB2IndexManager.java
@@ -18,27 +18,16 @@
 package org.jkiss.dbeaver.ext.db2.manager;
 
 import org.jkiss.code.Nullable;
-import org.jkiss.dbeaver.ext.db2.DB2Messages;
 import org.jkiss.dbeaver.ext.db2.model.DB2Index;
-import org.jkiss.dbeaver.ext.db2.model.DB2IndexColumn;
 import org.jkiss.dbeaver.ext.db2.model.DB2TableBase;
-import org.jkiss.dbeaver.ext.db2.model.DB2TableColumn;
-import org.jkiss.dbeaver.ext.db2.model.dict.DB2IndexType;
 import org.jkiss.dbeaver.ext.db2.model.dict.DB2UniqueRule;
 import org.jkiss.dbeaver.model.edit.DBECommandContext;
-import org.jkiss.dbeaver.model.impl.DBObjectNameCaseTransformer;
 import org.jkiss.dbeaver.model.impl.sql.edit.struct.SQLIndexManager;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
-import org.jkiss.dbeaver.model.struct.DBSEntityAttribute;
 import org.jkiss.dbeaver.model.struct.DBSObject;
 import org.jkiss.dbeaver.model.struct.cache.DBSObjectCache;
 import org.jkiss.dbeaver.model.struct.rdb.DBSIndexType;
-import org.jkiss.dbeaver.ui.UITask;
-import org.jkiss.dbeaver.ui.editors.object.struct.EditIndexPage;
-import org.jkiss.utils.CommonUtils;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -47,19 +36,6 @@ import java.util.Map;
  * @author Denis Forveille
  */
 public class DB2IndexManager extends SQLIndexManager<DB2Index, DB2TableBase> {
-
-    private static final String CONS_IX_NAME = "%s_%s_IDX";
-
-    private static final List<DBSIndexType> IX_TYPES;
-
-    static {
-        IX_TYPES = new ArrayList<>(DB2IndexType.values().length);
-        for (DB2IndexType db2IndexType : DB2IndexType.values()) {
-            if (db2IndexType.isValidForCreation()) {
-                IX_TYPES.add(db2IndexType.getDBSIndexType());
-            }
-        }
-    }
 
     @Override
     public boolean canEditObject(DB2Index object) {
@@ -75,42 +51,11 @@ public class DB2IndexManager extends SQLIndexManager<DB2Index, DB2TableBase> {
     @Override
     protected DB2Index createDatabaseObject(DBRProgressMonitor monitor, DBECommandContext context, final Object container, Object from, Map<String, Object> options) {
         DB2TableBase table = (DB2TableBase) container;
-        DB2Index index = new DB2Index(
+        return new DB2Index(
             (DB2TableBase) table,
             "INDEX",
             DBSIndexType.UNKNOWN,
             DB2UniqueRule.U);
-
-        return new UITask<DB2Index>() {
-            @Override
-            protected DB2Index runTask() {
-                EditIndexPage editPage = new EditIndexPage(
-                    DB2Messages.edit_db2_index_manager_dialog_title, index, IX_TYPES);
-                if (!editPage.edit()) {
-                    return null;
-                }
-
-                String tableName = CommonUtils.escapeIdentifier(table.getName());
-
-                String colName = CommonUtils.escapeIdentifier(editPage.getSelectedAttributes().iterator().next().getName());
-                String indexBaseName = String.format(CONS_IX_NAME, tableName, colName);
-                String indexName = DBObjectNameCaseTransformer.transformName(table.getDataSource(), indexBaseName);
-
-                index.setName(indexName);
-                index.setIndexType(editPage.getIndexType());
-                index.setUniqueRule(editPage.isUnique() ? DB2UniqueRule.U : DB2UniqueRule.D);
-
-                int colIndex = 1;
-                for (DBSEntityAttribute tableColumn : editPage.getSelectedAttributes()) {
-                    index.addColumn(new DB2IndexColumn(
-                        index,
-                        (DB2TableColumn) tableColumn,
-                        colIndex++,
-                        !Boolean.TRUE.equals(editPage.getAttributeProperty(tableColumn, EditIndexPage.PROP_DESC))));
-                }
-                return index;
-            }
-        }.execute();
     }
 
 }
