+++ plugins/org.jkiss.dbeaver.ext.db2/src/org/jkiss/dbeaver/ext/db2/manager/DB2UniqueKeyManager.java
@@ -18,24 +18,16 @@
 package org.jkiss.dbeaver.ext.db2.manager;
 
 import org.jkiss.code.Nullable;
-import org.jkiss.dbeaver.ext.db2.DB2Messages;
 import org.jkiss.dbeaver.ext.db2.model.DB2Table;
-import org.jkiss.dbeaver.ext.db2.model.DB2TableColumn;
-import org.jkiss.dbeaver.ext.db2.model.DB2TableKeyColumn;
 import org.jkiss.dbeaver.ext.db2.model.DB2TableUniqueKey;
 import org.jkiss.dbeaver.model.DBPEvaluationContext;
 import org.jkiss.dbeaver.model.edit.DBECommandContext;
 import org.jkiss.dbeaver.model.impl.sql.edit.struct.SQLConstraintManager;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
-import org.jkiss.dbeaver.model.struct.DBSEntityAttribute;
 import org.jkiss.dbeaver.model.struct.DBSEntityConstraintType;
 import org.jkiss.dbeaver.model.struct.DBSObject;
 import org.jkiss.dbeaver.model.struct.cache.DBSObjectCache;
-import org.jkiss.dbeaver.ui.UITask;
-import org.jkiss.dbeaver.ui.editors.object.struct.EditConstraintPage;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -47,9 +39,6 @@ public class DB2UniqueKeyManager extends SQLConstraintManager<DB2TableUniqueKey,
 
     private static final String                    SQL_DROP_PK = "ALTER TABLE %s DROP PRIMARY KEY ";
     private static final String                    SQL_DROP_UK = "ALTER TABLE %s DROP UNIQUE %s";
-
-    private static final DBSEntityConstraintType[] CONS_TYPES  = { DBSEntityConstraintType.PRIMARY_KEY,
-        DBSEntityConstraintType.UNIQUE_KEY };
 
     // -----------------
     // Business Contract
@@ -76,33 +65,7 @@ public class DB2UniqueKeyManager extends SQLConstraintManager<DB2TableUniqueKey,
     public DB2TableUniqueKey createDatabaseObject(DBRProgressMonitor monitor, DBECommandContext context, final Object table,
                                                   Object from, Map<String, Object> options)
     {
-        DB2TableUniqueKey constraint = new DB2TableUniqueKey((DB2Table) table, DBSEntityConstraintType.UNIQUE_KEY);
-
-        return new UITask<DB2TableUniqueKey>() {
-            @Override
-            protected DB2TableUniqueKey runTask()
-            {
-                EditConstraintPage editPage = new EditConstraintPage(DB2Messages.edit_db2_constraint_manager_dialog_title, constraint, CONS_TYPES);
-
-                if (!editPage.edit()) {
-                    return null;
-                }
-
-                constraint.setConstraintType(editPage.getConstraintType());
-                constraint.setName(editPage.getConstraintName());
-
-                List<DB2TableKeyColumn> columns = new ArrayList<>(editPage.getSelectedAttributes().size());
-                DB2TableKeyColumn column;
-                int colIndex = 1;
-                for (DBSEntityAttribute tableColumn : editPage.getSelectedAttributes()) {
-                    column = new DB2TableKeyColumn(constraint, (DB2TableColumn) tableColumn, colIndex++);
-                    columns.add(column);
-                }
-                constraint.setColumns(columns);
-
-                return constraint;
-            }
-        }.execute();
+        return new DB2TableUniqueKey((DB2Table) table, DBSEntityConstraintType.UNIQUE_KEY);
     }
 
     // ------
